refactor: #34 - 공통 UI DSKit으로 분리
