+++ Projects/Modules/DSKit/Sources/Component/NDGLOnGoingView.swift
@@ -1,15 +1,18 @@
 //
-//  HomeBannerOnGoingView.swift
-//  HomeFeature
+//  NDGLOnGoingView.swift
+//  DSKit
 //
-//  Created by 최안용 on 2/3/26.
+//  Created by kimnahun on 2026-02-22.
 //  Copyright © 2026 NDGL-iOS. All rights reserved.
 //
 
 import UIKit
 
-import DSKit
-final class HomeBannerOnGoingView: UIView {
+import Kingfisher
+import SnapKit
+import Then
+
+public final class NDGLOnGoingView: UIView {
     private let titleLabel = UILabel()
     private let dateLabel = UILabel()
     private let iconImageView = UIImageView()
@@ -24,20 +27,20 @@ final class HomeBannerOnGoingView: UIView {
     private let routeStackView = UIStackView()
     private let routeCardView = UIView()
     private let containerStackView = UIStackView()
-    
-    override init(frame: CGRect) {
+
+    override public init(frame: CGRect) {
         super.init(frame: frame)
-        
+
         setStyle()
         setUI()
         setLayout()
     }
-    
+
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-    
-    func configure(
+
+    public func configure(
         title: String,
         date: String,
         transportIcon: UIImage?,
@@ -52,15 +55,15 @@ final class HomeBannerOnGoingView: UIView {
         transportLabel.setText(.bodySM, text: transport, color: DSKitAsset.Colors.black400.color)
         durationLabel.setText(.bodySM, text: "\(duration) 체류 예상", color: DSKitAsset.Colors.black400.color)
         placeLabel.setText(.bodyLSB, text: place, color: DSKitAsset.Colors.black900.color)
-        
+
         if let url = URL(string: imageUrl) {
             imageView.kf.setImage(with: url)
         } else {
             imageView.backgroundColor = .systemGray5
         }
     }
-    
-    func prepareForReuse() {
+
+    public func prepareForReuse() {
         imageView.kf.cancelDownloadTask()
         titleLabel.text = nil
         dateLabel.text = nil
@@ -72,61 +75,61 @@ final class HomeBannerOnGoingView: UIView {
     }
 }
 
-private extension HomeBannerOnGoingView {
+private extension NDGLOnGoingView {
     func setStyle() {
         backgroundColor = .clear
-        
+
         dotLabel.do {
             $0.setText(.bodyMM, text: "•", color: DSKitAsset.Colors.black400.color)
         }
-        
+
         iconImageView.do {
             $0.tintColor = DSKitAsset.Colors.black500.color
         }
-        
+
         imageView.do {
             $0.layer.cornerRadius = 4.adjustedH
             $0.backgroundColor = .systemGray6
             $0.clipsToBounds = true
         }
-        
+
         titleStackView.do {
             $0.axis = .vertical
             $0.spacing = 4.adjustedH
             $0.alignment = .leading
         }
-        
+
         subInfoStackView.do {
             $0.axis = .horizontal
             $0.spacing = 4.adjusted
             $0.alignment = .center
         }
-        
+
         infoStackView.do {
             $0.axis = .vertical
             $0.spacing = 10.adjustedH
             $0.alignment = .leading
         }
-        
+
         routeStackView.do {
             $0.axis = .horizontal
             $0.spacing = 12.adjusted
             $0.alignment = .center
         }
-        
+
         routeCardView.do {
             $0.backgroundColor = DSKitAsset.Colors.white.color
             $0.layer.cornerRadius = 16.adjustedH
             $0.clipsToBounds = true
         }
-        
+
         containerStackView.do {
             $0.axis = .vertical
             $0.spacing = 16.adjustedH
             $0.alignment = .leading
         }
     }
-    
+
     func setUI() {
         titleStackView.addArrangedSubviews(titleLabel, dateLabel)
         subInfoStackView.addArrangedSubviews(iconImageView, transportLabel, dotLabel, durationLabel)
@@ -136,30 +139,30 @@ private extension HomeBannerOnGoingView {
         containerStackView.addArrangedSubviews(titleStackView, routeCardView)
         addSubview(containerStackView)
     }
-    
+
     func setLayout() {
         iconImageView.snp.makeConstraints {
             $0.size.equalTo(14.adjustedH)
         }
-        
+
         imageView.snp.makeConstraints {
             $0.size.equalTo(56.adjustedH)
         }
-        
+
         routeStackView.snp.makeConstraints {
             $0.directionalHorizontalEdges.equalToSuperview().inset(16.adjusted)
             $0.top.equalToSuperview().inset(12.adjustedH)
             $0.bottom.equalToSuperview().inset(16.adjustedH)
         }
-        
+
         routeCardView.snp.makeConstraints {
             $0.width.equalToSuperview()
         }
-        
+
         titleStackView.snp.makeConstraints {
             $0.directionalHorizontalEdges.equalToSuperview()
         }
-        
+
         containerStackView.snp.makeConstraints {
             $0.directionalHorizontalEdges.equalToSuperview().inset(16.adjusted)
             $0.top.equalToSuperview().inset(16.adjustedH)
